+++ packages/acs-ui-common/src/common.ts
@@ -1,10 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { _FileUploadCardsStrings } from '../FileUploadCards';
-/* @conditional-compile-remove(file-sharing) */
-import { useLocale } from '../../localization';
-
 /**
  * @internal
  * Converts units of rem to units of pixels
@@ -30,18 +26,4 @@ export const _preventDismissOnEvent = (
   ev: Event | React.FocusEvent | React.KeyboardEvent | React.MouseEvent
 ): boolean => {
   return ev.type === 'resize' || ev.type === 'scroll';
-};
-
-/**
- * Conditionally modify locale strings passed to the file card
- * @returns file upload card strings
- */
-export const useLocaleFileCardStringsTrampoline = (): _FileUploadCardsStrings => {
-  /* @conditional-compile-remove(file-sharing) */
-  return useLocale().strings.sendBox;
-  return {
-    removeFile: '',
-    uploadCompleted: '',
-    uploading: ''
-  };
 };
